requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
pillow==10.3.0
chromadb>=0.5.3
sentence-transformers==3.0.0
python-dotenv==1.0.1
aiosqlite==0.20.0
greenlet==3.0.3

# stage 2
anthropic==0.28.0
openai==1.30.0
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
numpy>=2.0.0
httpx==0.27.0
fpdf2>=2.7.9
asyncpg==0.29.0
